# Reworked signature scanner and memory service to reduce memory usage