… Anamnesis/Core/Memory/SignatureScanner.cs