_Note: This is the first of a series of optimization changes that I would like to propose to be introduced to Anamnesis. This one in particular tackles an issue I identified while running the application through a memory profiler._

---

### The issue

During application initialization, the signature scanner runs through the `.text` memory region of the game process, getting all of its data. During this scanning process, a temporary byte buffer is created, equal to the size of the entire `.text` memory region. Several of those signature scans run in parallel to load in actor, weather, terrain, action timeline, etc. During this time, different threads are loading in the game data from Lumina.

https://github.com/imchillin/Anamnesis/blob/dd42f7225d1dd8c06ab8bf25d34d1e0c09e0f918/Anamnesis/Core/Memory/SignatureScanner.cs#L129

The temporary buffers for the scan eventually get deallocated by the garbage collector (GC), shortly after the application startup. Whereas, the information retrieved from Lumina will persist, fragmented in sections within the Large Object Heap (LOH).

As visualized by JetBrains dotMemory, the LOH skyrockets to ~860 MB:
> Debug Build
![image](https://github.com/user-attachments/assets/f86d0bf3-2c83-4d99-918a-f9fa5c7d6286)

Upon digging deeper, the profiler reveals that the LOH is not only fragmented but also left mostly unused:
> Debug Build
![image](https://github.com/user-attachments/assets/8a43b3cb-a85f-4d49-b8a2-5019b5d62f92)

As a result, the Anamnesis application ends up taking 1.0+ GB of memory out of the user's PC:
> Release Build*
![image](https://github.com/user-attachments/assets/b5505abf-18b8-4c57-be97-50010dc2893c)

---

### The solution

By default, the GC never compacts LOH. Running such a process introduces a significant performance penalty, hence why its usage is generally frowned upon. Furthermore, manually using the GC does not resolve the issue of the initial LOH spike, which would cause Anamnesis to run out of space on a PC with limited RAM. ([Microsoft Docs](https://learn.microsoft.com/en-us/dotnet/api/system.runtime.gcsettings.largeobjectheapcompactionmode?view=net-6.0))

This is why I decided to rework the signature scanner, alongside some modifications to the memory service, to better handle memory read and write operations.

I tried two approaches for scanner:
1) Implementing the [Knuth–Morris–Pratt algorithm](https://en.wikipedia.org/wiki/Knuth%E2%80%93Morris%E2%80%93Pratt_algorithm).
2) Implementing the [Boyer–Moore string-search algorithm](https://en.wikipedia.org/wiki/Boyer%E2%80%93Moore_string-search_algorithm).

It seems that out of the two, the BM algorithm performed better for this use case, so I decided to use that one. Dalamud also seems to use this approach so opted to use their utility function to build the bad character shift table. The tricky part to creating the shift table is accounting for the needle bitmask and their solution is an optimal one.

However, this alone significantly reduces performance, increasing init times, compared to the old Anamnesis solution. This is because the `MemoryService.Read` (which uses ReadProcessMemory under the hood) function call is a relatively slow process (2000-3000ns per call). As a result, I combined the BM algorithm with a memory chunking process to reduce the number of memory read calls. The new algorithm also uses an overlap to account for signatures that might end up between two memory chunks.

After some testing, I opted for a **1kB** buffer chunk size. Increasing the size after this point seems to have a diminishing performance return, increasing memory usage.

---

### Results

The LOH size dropped from ~860 MB to just ~125 MB:
> Debug Build
![image](https://github.com/user-attachments/assets/941b7754-6b0a-4ece-a92d-bd097caf353a)

The LOH now sits between 85-90% utilization with a similar fragmentation percentage:
> Debug Build
![image](https://github.com/user-attachments/assets/7c69bff0-c422-467d-80a0-2a01c8f9d3b7)

After booting up the applicaiton, Anamnesis' memory usage is now:
> Release Build*
![image](https://github.com/user-attachments/assets/bafc26b2-bfbe-461f-b680-40fb30c85688)

---

### Memory vs Performance

Typically, reducing memory comes with a performance penalty. However, due to the use of the Boyer-Moore algorithm, the proposed solution not only tremendously decreases memory use but also improves the signature scanner's performance, thus decreasing initialization times.

On a **Release** build, the logs reveal the following initialization times:
**Before:**
> Initialized service: MemoryService in 1293ms
> Initialized service: AddressService in 139ms
> Services intialized in 3755ms

**After:**
> Initialized service: MemoryService in 678ms
> Initialized service: AddressService in 126ms
> Services intialized in 2383ms

\* The memory usage screenshots were taken directly after opening the application and switching to the Actor Tab (which defaults to the equipment editor view).