@StoiaCode 2024-08-30
Really well researched, thank you!

I've been made aware that Penumbra has implemented a basic signature scanner as well, but they used memory-mapped files instead of the memory chunking method you employed. I'm not sure which approach is more effective, and I'm fine with accepting your pull request as it stands. However, if you're interested in testing their approach against yours, I'd really appreciate it! Here's the Penumbra code that demonstrates their method: https://github.com/xivdev/Penumbra/blob/ff3e5410aac9e23606317e179f6278e710cb11ee/Penumbra/Interop/Hooks/ResourceLoading/PeSigScanner.cs#L26.