@ergoxiv 2024-08-30
Oooh, interesting. I didn't know Penumbra have their modified implementation of the scanner. I'll take a look and report back ^^