@ergoxiv 2024-08-30
I ran a bunch of tests using a direct copy of Penumbra's signature scanner.

That implementation seems to perform about the same memory-wise (within the margin of error) compared to the memory chunking reading approach:

![image](https://github.com/user-attachments/assets/58c3ef7e-2953-4305-926a-536d9a7e00ef)
![image](https://github.com/user-attachments/assets/99c4044c-8207-4cf5-83df-8a16828a7a12)
![image](https://github.com/user-attachments/assets/b02c6a5c-c1b1-44b8-8abf-9028b41c45fd)

In terms of speed, the logs show:
> Initialized service: MemoryService in 806ms _(though it hovers around 630-680ms on sequential app runs)_
> Initialized service: AddressService in 61ms
> Services intialized in 2468ms

Noticeably, the address service (the service runs the scans during its initialization) seems to initialize consistently about twice as fast, so it seems that it runs faster.

The only downside of Penumbra's implementation is the use of a third-party dependency (PeNet). Other than that, it is a really good solution.