@StoiaCode 2024-08-30
You sure you didnt accidentally copy paste the same screenshot, those numbers are almost too close 😆 

Neat, thanks for testing tho! Very interesting result. Let me know if you want/need to adjust anything, ill test it tonight and see if theres anything to complain about :)